Keep mask values as uint8_t rather than int.
Partial change, up to the point where I hit a compiler error, which I want
to deal with in a separate commit.

+++ src/dials/algorithms/background/simple/mosflm_outlier_rejector.h
@@ -51,7 +51,7 @@ namespace dials { namespace algorithms { namespace background {
      * @params mask The shoebox mask
      */
     virtual void mark(const af::const_ref<double, af::c_grid<3> >& data,
-                      af::ref<int, af::c_grid<3> > mask) const {
+                      af::ref<uint8_t, af::c_grid<3> > mask) const {
       // Check the input
       DIALS_ASSERT(data.accessor().all_eq(mask.accessor()));
 
@@ -61,7 +61,7 @@ namespace dials { namespace algorithms { namespace background {
       for (std::size_t i = 0; i < data.accessor()[0]; ++i) {
         // Get the 2D slices
         af::const_ref<double, af::c_grid<2> > data_2d(&data[i * xysize], accessor);
-        af::ref<int, af::c_grid<2> > mask_2d(&mask[i * xysize], accessor);
+        af::ref<uint8_t, af::c_grid<2> > mask_2d(&mask[i * xysize], accessor);
 
         // Compute the initial mask using a subset of the available pixels
         compute_initial_mask(data_2d, mask_2d);
@@ -98,8 +98,8 @@ namespace dials { namespace algorithms { namespace background {
      * intensity and then update the mask for those pixels.
      */
     void compute_initial_mask(const af::const_ref<double, af::c_grid<2> >& data,
-                              af::ref<int, af::c_grid<2> > mask) const {
-      int code = Valid | Background;
+                              af::ref<uint8_t, af::c_grid<2> > mask) const {
+      uint8_t code = Valid | Background;
       std::vector<std::size_t> index;
       index.reserve(data.size());
       for (std::size_t i = 0; i < mask.size(); ++i) {
@@ -122,7 +122,7 @@ namespace dials { namespace algorithms { namespace background {
      * Compute the background plane given shoebox data and mask
      */
     void compute_background(const af::const_ref<double, af::c_grid<2> >& data,
-                            const af::const_ref<int, af::c_grid<2> >& mask,
+                            const af::const_ref<uint8_t, af::c_grid<2> >& mask,
                             double& a,
                             double& b,
                             double& c) const {
@@ -182,7 +182,7 @@ namespace dials { namespace algorithms { namespace background {
      * deviation of each point from the plane.
      */
     void compute_final_mask(const af::const_ref<double, af::c_grid<2> >& data,
-                            af::ref<int, af::c_grid<2> > mask,
+                            af::ref<uint8_t, af::c_grid<2> > mask,
                             double a,
                             double b,
                             double c) const {
